requirements.txt
cryptography
numpy
pandas
PyAutoGUI
pydal
pydantic
pymysql
python-dotenv
ttkwidgets
